requirements.txt
feedparser
openai
tweepy 
python-dotenv
crawl4ai
iointel
tinydb
